move worker-go to worker.
