--- worker-go/.env
@@ -1,14 +1,0 @@
-DATABASE_URL="postgresql://postgres:postgres@localhost:5432/mydb"
-
-ALIENTVAULT_KEY=test
-BINARYEDGE_KEY=test
-CHAOS_KEY=test
-CLOUDFLARE_KEY=test
-HUNTER_KEY=test
-IPINFO_KEY=test
-NETWORKSDB_KEY=test
-SHODAN_KEY=test
-URLSCAN_KEY=test
-VIRUSTOTAL_KEY=test
-WHOISXMLAPI_KEY=test
-SECURITYTRAILS_KEY=test
