--- worker-go/.gitignore
@@ -1,3 +1,0 @@
-.env.local
-tmp
-resume.cfg
