--- worker-go/db/.gitignore
@@ -1,2 +1,0 @@
-# gitignore generated by Prisma Client Go. DO NOT EDIT.
-*_gen.go
