--- worker/.dockerignore
@@ -1,2 +1,0 @@
-target
-.env
