+++ worker/.env
@@ -1,5 +1,4 @@
 DATABASE_URL="postgresql://postgres:postgres@localhost:5432/mydb"
-SQLX_OFFLINE=true
 
 ALIENTVAULT_KEY=test
 BINARYEDGE_KEY=test
