+++ worker/.gitignore
@@ -1,9 +1,3 @@
-target
-amass/amass-old.json
-amass/amass.json
-amass/amass.log
-amass/amass.txt
-amass/config.ini
-amass/indexes.bolt
-
 .env.local
+tmp
+resume.cfg
