--- worker/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "worker"
-version = "0.1.0"
-authors = ["Chrisando Pramudhita <chrsep@protonmail.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tokio = { version = "1.12.0", features = ["full"] }
-dotenv = "0.15.0"
-sqlx = { version = "0.5", features = [ "runtime-tokio-rustls", "postgres", "macros", "chrono", "offline" ] }
-env_logger = { version = "0.9.0" }
-log = "0.4.14"
-serde_json = "1.0"
-serde = "1.0"
-envsubst = "0.2.0"
