--- worker/Dockerfile
@@ -1,13 +1,0 @@
-FROM rust:1.55 AS builder
-WORKDIR /app
-COPY . .
-RUN cargo install -v --path .
-
-FROM gcr.io/distroless/cc
-WORKDIR /app
-
-COPY --from=builder /usr/local/cargo/bin/worker /app/worker
-COPY --from=builder /app/wordlist /app/wordlist
-COPY --from=builder /app/amass /app/amass
-
-CMD ["./worker"]
