+++ worker/db/.gitignore
@@ -1,0 +1,2 @@
+# gitignore generated by Prisma Client Go. DO NOT EDIT.
+*_gen.go
