+++ worker/go.mod
@@ -1,0 +1,41 @@
+module worker-go
+
+go 1.17
+
+require (
+	github.com/iancoleman/strcase v0.0.0-20190422225806-e506e3ef7365
+	github.com/joho/godotenv v1.4.0
+	github.com/prisma/prisma-client-go v0.12.2
+	github.com/projectdiscovery/subfinder/v2 v2.4.9
+	github.com/shopspring/decimal v1.2.0
+	github.com/takuoki/gocase v1.0.0
+)
+
+require (
+	github.com/andres-erbsen/clock v0.0.0-20160526145045-9e14626cd129 // indirect
+	github.com/corpix/uarand v0.1.1 // indirect
+	github.com/hako/durafmt v0.0.0-20210316092057-3a2c319c1acd // indirect
+	github.com/json-iterator/go v1.1.11 // indirect
+	github.com/lib/pq v1.10.0 // indirect
+	github.com/logrusorgru/aurora v2.0.3+incompatible // indirect
+	github.com/miekg/dns v1.1.43 // indirect
+	github.com/mitchellh/mapstructure v1.4.1 // indirect
+	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
+	github.com/modern-go/reflect2 v1.0.1 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
+	github.com/projectdiscovery/chaos-client v0.1.8 // indirect
+	github.com/projectdiscovery/dnsx v1.0.3 // indirect
+	github.com/projectdiscovery/gologger v1.1.4 // indirect
+	github.com/projectdiscovery/retryabledns v1.0.12 // indirect
+	github.com/rs/xid v1.3.0 // indirect
+	github.com/spyse-com/go-spyse v1.2.3 // indirect
+	github.com/tomnomnom/linkheader v0.0.0-20180905144013-02ca5825eb80 // indirect
+	go.uber.org/atomic v1.9.0 // indirect
+	go.uber.org/multierr v1.7.0 // indirect
+	go.uber.org/ratelimit v0.2.0 // indirect
+	go.uber.org/zap v1.19.1 // indirect
+	golang.org/x/net v0.0.0-20210813160813-60bc85c4be6d // indirect
+	golang.org/x/sys v0.0.0-20211113001501-0c823b97ae02 // indirect
+	golang.org/x/time v0.0.0-20210723032227-1f47c861a9ac // indirect
+	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
+)
