+++ worker/schema.prisma
@@ -1,0 +1,79 @@
+// This is your Prisma schema file,
+// learn more about it in the docs: https://pris.ly/d/prisma-schema
+
+datasource db {
+  provider = "postgresql"
+  url      = env("DATABASE_URL")
+}
+
+generator client {
+    provider = "go run github.com/prisma/prisma-client-go"
+}
+
+model Company {
+  id         Int      @id @default(autoincrement())
+  name       String
+  bountyLink String   @default("")
+  icon       String   @default("")
+  domains    Domain[]
+}
+
+model Domain {
+  name      String   @id
+  confirmed Boolean  @default(false)
+  createdAt DateTime @default(now())
+  updatedAt DateTime @default(now())
+
+  company        Company   @relation(fields: [companyId], references: [id], onDelete: Cascade)
+  companyId      Int
+  rootDomain     Domain?   @relation("DomainToSubDomain", fields: [rootDomainName], references: [name], onDelete: Cascade)
+  rootDomainName String?
+  subDomains     Domain[]  @relation("DomainToSubDomain")
+  services       Service[]
+}
+
+model Service {
+  port        Int
+  name        String
+  description String
+  createdAt   DateTime @default(now())
+  updatedAt   DateTime @default(now())
+
+  cnameRecords String[]
+  aRecords     String[]
+
+  domain        Domain          @relation(fields: [domainName], references: [name], onDelete: Cascade)
+  domainName    String
+  probeResponse ProbeResponse[]
+
+  @@id([port, domainName])
+}
+
+model ProbeResponse {
+  bodySHA String @id
+
+  url       String
+  host      String
+  method    String
+  scheme    String
+  webserver String
+  timestamp DateTime
+
+  title        String
+  header       String
+  body         String
+  statusCode   Int
+  contentType  String
+  path         String
+  responseTime String
+
+  service           Service @relation(fields: [servicePort, serviceDomainName], references: [port, domainName])
+  servicePort       Int
+  serviceDomainName String
+  techs             Tech[]
+}
+
+model Tech {
+  name          String @id
+  probeResponse ProbeResponse[]
+}
