--- worker/sqlx-data.json
@@ -1,120 +1,0 @@
-{
-  "db": "PostgreSQL",
-  "a1b3bb3d6a2266c43900a7990262e2fae6b46da31b981bcbb06b19ad974cdb58": {
-    "query": "\n                insert into \"IpAddress\"(\"subDomainDomain\", ip, cidr, asn, \"desc\", \"updatedAt\")\n                values ($1, $2, $3, $4, $5, now())\n                on conflict do nothing\n                ",
-    "describe": {
-      "columns": [],
-      "parameters": {
-        "Left": [
-          "Text",
-          "Text",
-          "Text",
-          "Int4",
-          "Text"
-        ]
-      },
-      "nullable": []
-    }
-  },
-  "a3be716bbc11e4bd5308926f0c7980a101e2e1626a75199daa4ed3cca6db13e5": {
-    "query": "\n            insert into \"SubDomain\"(domain, \"amassTag\",  \"rootDomainDomain\", sources, \"updatedAt\")\n            values ($1, $2, $3, $4, now())\n            on conflict do nothing\n            ",
-    "describe": {
-      "columns": [],
-      "parameters": {
-        "Left": [
-          "Text",
-          "Text",
-          "Text",
-          "TextArray"
-        ]
-      },
-      "nullable": []
-    }
-  },
-  "b9ccfcc0a4eb50d714d8ccf2dd61cfbfc9417a9f9b7aa84be360f5cb99f0a439": {
-    "query": "\n            SELECT *  from \"RootDomain\" \n            where confirmed and now() -\"lastDNSRecon\"  > interval '5 hour' ",
-    "describe": {
-      "columns": [
-        {
-          "ordinal": 0,
-          "name": "domain",
-          "type_info": "Text"
-        },
-        {
-          "ordinal": 1,
-          "name": "companyId",
-          "type_info": "Int4"
-        },
-        {
-          "ordinal": 2,
-          "name": "confirmed",
-          "type_info": "Bool"
-        },
-        {
-          "ordinal": 3,
-          "name": "createdAt",
-          "type_info": "Timestamp"
-        },
-        {
-          "ordinal": 4,
-          "name": "lastDNSRecon",
-          "type_info": "Timestamp"
-        },
-        {
-          "ordinal": 5,
-          "name": "isDNSReconRunning",
-          "type_info": "Bool"
-        }
-      ],
-      "parameters": {
-        "Left": []
-      },
-      "nullable": [
-        false,
-        false,
-        false,
-        false,
-        false,
-        false
-      ]
-    }
-  },
-  "b9f81d45fcb25deccab6add40663e54f8c8ce287aa315c9dd84cd0dc94017ec7": {
-    "query": "\n            insert into \"RootDomain\" (domain, \"companyId\")\n            select * from unnest($1::text[], $2::int[])\n            on conflict do nothing\n            ",
-    "describe": {
-      "columns": [],
-      "parameters": {
-        "Left": [
-          "TextArray",
-          "Int4Array"
-        ]
-      },
-      "nullable": []
-    }
-  },
-  "cf36c5b48f9bda30380ac4a2ec52c21406728f32fa173f6ab7a6cedfd0d0f312": {
-    "query": "\n            update \"RootDomain\"\n            set \"isDNSReconRunning\" = $1::bool\n            where \"domain\" = $2::text\n            ",
-    "describe": {
-      "columns": [],
-      "parameters": {
-        "Left": [
-          "Bool",
-          "Text"
-        ]
-      },
-      "nullable": []
-    }
-  },
-  "f36cddb3e5cc81d643d5d303cac3e8a22d899d03bef81754049a1b73a3939d0b": {
-    "query": "\n            update \"RootDomain\"\n            set \"lastDNSRecon\" = now()\n            where \"domain\" = $1::text\n            ",
-    "describe": {
-      "columns": [],
-      "parameters": {
-        "Left": [
-          "Text"
-        ]
-      },
-      "nullable": []
-    }
-  }
-}
