--- worker/test_fixtures/amass.json
@@ -1,6 +1,0 @@
-{"name":"sl.owasp.org","domain":"owasp.org","addresses":[{"ip":"172.67.10.39","cidr":"172.67.0.0/16","asn":13335,"desc":"CLOUDFLARENET - Cloudflare, Inc."}],"tag":"api","sources":["HackerTarget"]}
-{"name":"kerala.owasp.org","domain":"owasp.org","addresses":[{"ip":"172.67.10.39","cidr":"172.67.0.0/16","asn":13335,"desc":"CLOUDFLARENET - Cloudflare, Inc."},{"ip":"2606:4700:10::6816:1a4d","cidr":"2606:4700:10::/44","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."}],"tag":"api","sources":["HackerTarget"]}
-{"name":"www.owasp.org","domain":"owasp.org","addresses":[{"ip":"2606:4700:10::6816:1a4d","cidr":"2606:4700:10::/44","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."},{"ip":"2606:4700:10::6816:1b4d","cidr":"2606:4700:10::/44","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."},{"ip":"104.22.27.77","cidr":"104.22.16.0/20","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."},{"ip":"104.22.26.77","cidr":"104.22.16.0/20","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."},{"ip":"172.67.10.39","cidr":"172.67.0.0/16","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."},{"ip":"2606:4700:10::ac43:a27","cidr":"2606:4700:10::/44","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."}],"tag":"scrape","sources":["Bing"]}
-{"name":"austin.owasp.org","domain":"owasp.org","addresses":[{"ip":"172.67.10.39","cidr":"172.67.0.0/16","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."},{"ip":"104.22.27.77","cidr":"104.22.16.0/20","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."},{"ip":"104.22.26.77","cidr":"104.22.16.0/20","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."},{"ip":"2606:4700:10::ac43:a27","cidr":"2606:4700:10::/44","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."},{"ip":"2606:4700:10::6816:1b4d","cidr":"2606:4700:10::/44","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."},{"ip":"2606:4700:10::6816:1a4d","cidr":"2606:4700:10::/44","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."}],"tag":"api","sources":["HackerTarget"]}
-{"name":"owasp.org","domain":"owasp.org","addresses":[{"ip":"2606:4700:10::6816:1b4d","cidr":"2606:4700:10::/44","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."},{"ip":"2606:4700:10::6816:1a4d","cidr":"2606:4700:10::/44","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."},{"ip":"172.67.10.39","cidr":"172.67.0.0/16","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."},{"ip":"104.22.27.77","cidr":"104.22.16.0/20","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."},{"ip":"104.22.26.77","cidr":"104.22.16.0/20","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."},{"ip":"2606:4700:10::ac43:a27","cidr":"2606:4700:10::/44","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."}],"tag":"cert","sources":["DNS","HackerTarget","SonarSearch","Bing","DuckDuckGo","Riddler","Yahoo","AlienVault","CertSpotter","Hunter","Baidu","GoogleCT","Mnemonic","SiteDossier","Ask","BinaryEdge","WhoisXMLAPI","BufferOver","Censys","RapidDNS","ThreatMiner","Alterations","ThreatCrowd","VirusTotal","Chaos","IPv4Info","CommonCrawl","Wayback","Crtsh","URLScan","ArchiveIt"]}
-{"name":"20thanniversary.owasp.org","domain":"owasp.org","addresses":[{"ip":"2606:4700:10::6816:1a4d","cidr":"2606:4700:10::/44","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."},{"ip":"172.67.10.39","cidr":"172.67.0.0/16","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."},{"ip":"104.22.27.77","cidr":"104.22.16.0/20","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."},{"ip":"104.22.26.77","cidr":"104.22.16.0/20","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."},{"ip":"2606:4700:10::ac43:a27","cidr":"2606:4700:10::/44","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."},{"ip":"2606:4700:10::6816:1b4d","cidr":"2606:4700:10::/44","asn":13335,"desc":"CLOUDFLARENET-AS CloudFlare, Inc. - CLOUDFLARENET-AS Cloudflare, Inc."}],"tag":"api","sources":["URLScan"]}
